Added player respawning
Player is reset to a checkpoint when killed, when this happens all enemies are destroyed and currently active waves are reset. Each checkpoint requires a collider to trigger its use when the player comes into contact with it. They also need a 'checkpoint number' that lets the system know if the checkpoint is further on in the level than the players last checkpoint.