Implemented ambient sounds
Ambient zones can be placed around the scene and used to determine which set of ambient sounds can be chosen. Sounds are given a weight to decide how likely they are to be picked. Sounds are played in a random direction to the player and the further they are from the ambient zone the quieter the sound will be.